backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==1.26.4
openai==1.78.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.5
pandas==2.2.3
pbr==6.1.1
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qiskit==2.0.0
requests==2.32.3
rsa==4.9.1
rustworkx==0.16.0
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stevedore==5.4.1
symengine==0.13.0
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
flask-cors==3.0.10
gunicorn==20.1.0
